Remove the "mount /boot" hack
So as it turns out, I had some old autofs configurations lying around
which were mounting something to /boot and conflicted with the ESP
partition that it *supposed* to be mounted there. This caused /boot
to show up empty without this hack, but removing the autofs config put
an end to this silliness.